TP2/requirements.txt
aiofiles
aiohttp
beautifulsoup4
lxml
html5lib
playwright==1.40.0
Pillow==10.1.0
requests==2.31.0